% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{costanza}
\alias{costanza}
\title{Multi-resolution Goodness-of-Fit}
\usage{
costanza(data, reference, resolution)
}
\arguments{
\item{data}{matrix of data.}

\item{reference}{matrix of data to be compared (must have the same size as `x`).}

\item{resolution}{desired number of clusters}
}
\value{
data frame with all multireolution parameters and metrics.
}
\description{
Compute multi-resolution Goodness-of-Fit on two matrices
with equal dimensions.
}
\examples{
require(raster)
require(ggplot2)
# Costanza with 10x10 images
f1 <- system.file("extdata/tif/costanza1.tif", package = "mrgf")
f2 <- system.file("extdata/tif/costanza2.tif", package = "mrgf")
# open files        
r1 <- raster::raster(f1)
r2 <- raster::raster(f2)
# plot images
plot(r1)
plot(r2)
# load data into matrices
m1 <- matrix(r1, nrow = nrow(r1))
m2 <- matrix(r2, nrow = nrow(r2))
# perform multi-resolution goodness-of-fit        
result <- costanza(m1, m2, resolution = 1:10)
# plot the result
ggplot(result) + 
    geom_line(aes(resolution, 
                  class_acc, 
                  colour = as.factor(classes), 
                  group = classes)) +
    geom_line(aes(resolution, global_acc))

}
\references{
Costanza_Paper_Reference
}
\author{
Rolf Simoes, \email{rolf.simoes@inpe.br}
}
